requirements.txt
selenium==4.14.0
pillow==10.1.0